# docs/openapi-indexer.yaml
openapi: 3.0.3
info:
  title: Agent Work Chain Indexer API
  version: 0.2.0
  description: Read-only discovery API for tasks, agents, events, and billing.
servers:
  - url: http://localhost:3000
paths:
...
  /v1/tasks:
    get:
      summary: List tasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [Created, InProgress, Completed, Disputed, Resolved, Cancelled]
        - in: query
          name: creator
          schema:
            type: string
        - in: query
          name: agent
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 0
            maximum: 200
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Task list
  /v1/tasks/{id}:
    get:
      summary: Get task detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task detail
        '404':
          description: Task not found
...